Rename plugin module with prefix
