+++ AGENTS.md
@@ -2,4 +2,5 @@
 
 * Gradle commands are time-consuming because they download large dependencies. Avoid running any Gradle commands (such as `gradle` or `./gradlew`) in this repository.
 * There are no mandatory programmatic checks for this repository.
-
+* Use common prefixes across all modules within multi-module projects. The prefix for this project is `move-tab`; modules should be named using this prefix (e.g. `move-tab-plugin`).
+* `AGENTS.md` should be continually updated when new or helpful standards or conventions are brought up during interactions.
