+++ move-tab-plugin/src/main/kotlin/com/mikejhill/intellij/movetab/actions/MoveTabRight.kt
