+++ move-tab-plugin/src/main/resources/META-INF/MANIFEST.MF
