+++ move-tab-plugin/src/main/resources/META-INF/plugin.xml
